// src/frontend/src/api/generated/types.gen.ts
export type ObjectAttributeDefinition = {
  target: "object";
  name: string;
  unit?: {
    /**
     * Unit symbol
     */
    symbol: string | null;
    /**
     * Long name of the unit
     */
    name: string | null;
    /**
     * The unit dimensionality, compatible with the `pint` python package
     */
    dim: {
      [key: string]: number;
    };
  } | null;
  objectType: string;
  dynamic: boolean;
};